sync mocks prototype
