+++ tests/e2e/runtime/signin.spec.ts
@@ -5,16 +5,9 @@ import { BrowserLaunchOptions } from "../../constants";
 import { SignInSocialWidget } from "../maps/signin-social";
 import { SignInBasicWidget } from "../maps/signin-basic";
 import { User } from "../../mocks";
+import setupMocks from "../mock";
 
-export async function signIn(page: puppeteer.Page, user: User): Promise<void> {
-    const config = await Utils.getConfig();
-    await Utils.mock(page, {
-        "/identity": {
-            id: config.signin.id
-        },
-        "/sso-refresh": "OK",
-    });
-
+export async function signIn(page: puppeteer.Page, config: any, user: User): Promise<void> {
     await page.goto(config.urls.signin);
 
     const signInWidget = new SignInBasicWidget(page);
@@ -30,13 +23,14 @@ describe("User sign-in flow", async () => {
         browser = await puppeteer.launch(BrowserLaunchOptions);
     });
     after(async () => {
-        browser.close();
+        await browser.close();
     });
 
     it("User can sign-in with basic credentials", async () => {
         const page = await browser.newPage();
         const confirmedUser = await Utils.getConfirmedUserBasic();
-        await signIn(page, confirmedUser);
+        await setupMocks(page, config, confirmedUser);
+        await signIn(page, config, confirmedUser);
         expect(page.url()).to.equal(config.urls.home);
     });
     /*
